+++ src/MoreIcons.hpp
@@ -20,7 +20,6 @@ struct ImageData {
     std::string name;
     std::string frameName;
     TexturePack pack;
-    IconType type;
     int index;
     bool blend;
     bool tint;
@@ -73,35 +72,8 @@ public:
     static inline LogType HIGHEST_SEVERITY = LogType::Info;
 
     static BS::thread_pool& sharedPool() {
-        static BS::thread_pool _sharedPool(std::thread::hardware_concurrency() - 2);
+        static BS::thread_pool _sharedPool(std::thread::hardware_concurrency());
         return _sharedPool;
-    }
-
-    static void clear() {
-        MoreIconsAPI::ICONS.clear();
-        ICON_INFO.clear();
-        MoreIconsAPI::SHIPS.clear();
-        SHIP_INFO.clear();
-        MoreIconsAPI::BALLS.clear();
-        BALL_INFO.clear();
-        MoreIconsAPI::UFOS.clear();
-        UFO_INFO.clear();
-        MoreIconsAPI::WAVES.clear();
-        WAVE_INFO.clear();
-        MoreIconsAPI::ROBOTS.clear();
-        ROBOT_INFO.clear();
-        MoreIconsAPI::SPIDERS.clear();
-        SPIDER_INFO.clear();
-        MoreIconsAPI::SWINGS.clear();
-        SWING_INFO.clear();
-        MoreIconsAPI::JETPACKS.clear();
-        JETPACK_INFO.clear();
-        MoreIconsAPI::TRAILS.clear();
-        saveTrails();
-        TRAIL_INFO.clear();
-        removeSaved();
-        LOGS.clear();
-        HIGHEST_SEVERITY = LogType::Info;
     }
 
     static void removeSaved() {
@@ -117,33 +89,20 @@ public:
         geode::Mod::get()->setSavedValue<std::vector<std::string>>("trails", {});
     }
 
-    static void restoreSaved() {
-        geode::Mod::get()->setSavedValue("icons", MoreIconsAPI::ICONS);
-        geode::Mod::get()->setSavedValue("ships", MoreIconsAPI::SHIPS);
-        geode::Mod::get()->setSavedValue("balls", MoreIconsAPI::BALLS);
-        geode::Mod::get()->setSavedValue("ufos", MoreIconsAPI::UFOS);
-        geode::Mod::get()->setSavedValue("waves", MoreIconsAPI::WAVES);
-        geode::Mod::get()->setSavedValue("robots", MoreIconsAPI::ROBOTS);
-        geode::Mod::get()->setSavedValue("spiders", MoreIconsAPI::SPIDERS);
-        geode::Mod::get()->setSavedValue("swings", MoreIconsAPI::SWINGS);
-        geode::Mod::get()->setSavedValue("jetpacks", MoreIconsAPI::JETPACKS);
-        geode::Mod::get()->setSavedValue("trails", MoreIconsAPI::TRAILS);
-    }
-
-    static void load(LoadingLayer*);
-
     static std::vector<std::filesystem::directory_entry> naturalSort(const std::filesystem::path& path);
 
     static void naturalSort(std::vector<std::string>& vec);
 
+    static bool naturalSorter(const std::string& aStr, const std::string& bStr);
+
     static std::vector<std::filesystem::path> getTexturePacks();
 
-    static std::string replaceEnd(const std::string& str, const std::string& end, const std::string& replace, bool check = false) {
-        return !check || str.ends_with(end) ? str.substr(0, str.size() - end.size()) + replace : str;
+    static std::string replaceEnd(const std::string& str, std::string_view end, std::string_view replace, bool check = false) {
+        return !check || str.ends_with(end) ? fmt::format("{}{}", str.substr(0, str.size() - end.size()), replace) : str;
     }
 
     static void loadIcons(
-        const std::vector<std::filesystem::path>& packs, const std::string& suffix, IconType type
+        const std::vector<std::filesystem::path>& packs, std::string_view suffix, IconType type
     );
 
     static void loadIcon(const std::filesystem::path& path, const TexturePack& pack, IconType type);
@@ -164,49 +123,25 @@ public:
     }
 
     static void changeSimplePlayer(SimplePlayer* player, IconType type) {
-        MoreIconsAPI::updateSimplePlayer(player, geode::Mod::get()->getSavedValue<std::string>(savedForType(type), ""), type);
+        auto sdi = geode::Loader::get()->getLoadedMod("weebify.separate_dual_icons");
+        MoreIconsAPI::updateSimplePlayer(player, MoreIconsAPI::activeForType(type, sdi && sdi->getSavedValue("2pselected", false)), type);
     }
 
     static void changeSimplePlayer(SimplePlayer* player, IconType type, bool dual) {
-        MoreIconsAPI::updateSimplePlayer(player, geode::Mod::get()->getSavedValue<std::string>(MoreIconsAPI::savedForType(type, dual), ""), type);
-    }
-
-    static bool doesExist(cocos2d::CCSpriteFrame* frame) {
-        return frame != nullptr && frame->getTag() != 105871529;
-    }
-
-    static std::string getDual(const std::string& name, bool dual) {
-        return geode::Loader::get()->isModLoaded("weebify.separate_dual_icons") && dual ? name + "-dual" : name;
-    }
-
-    static void swapDual(const std::string& name) {
-        auto dualName = name + "-dual";
-        auto normalIcon = geode::Mod::get()->getSavedValue<std::string>(name, "");
-        auto dualIcon = geode::Mod::get()->getSavedValue<std::string>(dualName, "");
-        geode::Mod::get()->setSavedValue(name, dualIcon);
-        geode::Mod::get()->setSavedValue(dualName, normalIcon);
+        MoreIconsAPI::updateSimplePlayer(player, MoreIconsAPI::activeForType(type, dual), type);
     }
 
     static std::unordered_map<std::string, TexturePack>& infoForType(IconType type) {
         switch (type) {
-            case IconType::Cube:
-                return ICON_INFO;
-            case IconType::Ship:
-                return SHIP_INFO;
-            case IconType::Ball:
-                return BALL_INFO;
-            case IconType::Ufo:
-                return UFO_INFO;
-            case IconType::Wave:
-                return WAVE_INFO;
-            case IconType::Robot:
-                return ROBOT_INFO;
-            case IconType::Spider:
-                return SPIDER_INFO;
-            case IconType::Swing:
-                return SWING_INFO;
-            case IconType::Jetpack:
-                return JETPACK_INFO;
+            case IconType::Cube: return ICON_INFO;
+            case IconType::Ship: return SHIP_INFO;
+            case IconType::Ball: return BALL_INFO;
+            case IconType::Ufo: return UFO_INFO;
+            case IconType::Wave: return WAVE_INFO;
+            case IconType::Robot: return ROBOT_INFO;
+            case IconType::Spider: return SPIDER_INFO;
+            case IconType::Swing: return SWING_INFO;
+            case IconType::Jetpack: return JETPACK_INFO;
             default: {
                 static std::unordered_map<std::string, TexturePack> empty;
                 return empty;
@@ -214,12 +149,7 @@ public:
         }
     }
 
-    static std::string savedForType(IconType type) {
-        auto sdi = geode::Loader::get()->getLoadedMod("weebify.separate_dual_icons");
-        return MoreIconsAPI::savedForType(type, sdi && sdi->getSavedValue("2pselected", false));
-    }
-
-    static std::string getFrameName(const std::string& name, const std::string prefix, IconType type) {
+    static std::string getFrameName(const std::string& name, const std::string& prefix, IconType type) {
         if (type != IconType::Robot && type != IconType::Spider) {
             if (name.ends_with("_2_001.png")) return fmt::format("{}_2_001.png"_spr, prefix);
             else if (name.ends_with("_3_001.png")) return fmt::format("{}_3_001.png"_spr, prefix);
@@ -240,37 +170,6 @@ public:
     static SimplePlayer* findPlayer(cocos2d::CCNode* node) {
         if (!node) return nullptr;
         return geode::cocos::findFirstChildRecursive<SimplePlayer>(node, [](auto) { return true; });
-    }
-
-    static std::vector<std::string> getPage(IconType type, int page) {
-        auto& vec = MoreIconsAPI::vectorForType(type);
-        auto customPage = page - (GameManager::get()->countForType(type) + 35) / 36;
-        if (vec.size() <= customPage * 36) return {};
-        if (customPage < 0) return {};
-        return std::vector<std::string>(vec.begin() + customPage * 36, vec.begin() + std::min((int)vec.size(), (customPage + 1) * 36));
-    }
-
-    static int countForType(IconType type) {
-        return MoreIconsAPI::vectorForType(type).size() + ((GameManager::get()->countForType(type) + 35) / 36 * 36);
-    }
-
-    static int wrapPage(IconType type, int page) {
-        auto pages = (countForType(type) + 35) / 36;
-        return pages > 0 ? page < 0 ? pages - 1 : page >= pages ? 0 : page : 0;
-    }
-
-    static int findIconPage(IconType type) {
-        auto active = (GameManager::get()->activeIconForType(type) - 1) / 36;
-        auto& vec = MoreIconsAPI::vectorForType(type);
-        auto savedType = savedForType(type);
-        if (savedType.empty()) return active;
-        auto it = std::find(vec.begin(), vec.end(), geode::Mod::get()->getSavedValue<std::string>(savedType, ""));
-        return it == vec.end() ? active : ((GameManager::get()->countForType(type) + 35) / 36) + (it - vec.begin()) / 36;
-    }
-
-    static bool isNormalPage(int page, IconType type) {
-        auto count = GameManager::get()->countForType(type);
-        return page * 36 < count;
     }
 
     static void showInfoPopup(bool folderButton = false) {