+++ src/api/MoreIconsAPI.cpp
@@ -21,7 +21,7 @@ std::vector<std::string>& MoreIconsAPI::vectorForType(IconType type) {
     }
 }
 
-std::string MoreIconsAPI::savedForType(IconType type, bool dual) {
+std::string_view MoreIconsAPI::savedForType(IconType type, bool dual) {
     auto isDual = Loader::get()->isModLoaded("weebify.separate_dual_icons") && dual;
     switch (type) {
         case IconType::Cube: return isDual ? "icon-dual" : "icon";
@@ -41,7 +41,7 @@ std::string MoreIconsAPI::savedForType(IconType type, bool dual) {
 void MoreIconsAPI::updateSimplePlayer(SimplePlayer* player, const std::string& icon, IconType type) {
     if (!player || icon.empty() || !hasIcon(icon, type)) return;
 
-    MoreIconsAPI::setUserObject(player, icon);
+    setUserObject(player, icon);
 
     if (type == IconType::Robot) {
         updateRobotSprite(player->m_robotSprite, icon, type);
@@ -81,7 +81,7 @@ void MoreIconsAPI::updateSimplePlayer(SimplePlayer* player, const std::string& i
 void MoreIconsAPI::updateRobotSprite(GJRobotSprite* sprite, const std::string& icon, IconType type) {
     if (!sprite || icon.empty() || !hasIcon(icon, type)) return;
 
-    MoreIconsAPI::setUserObject(sprite, icon);
+    setUserObject(sprite, icon);
 
     sprite->setBatchNode(nullptr);
     sprite->m_paSprite->setBatchNode(nullptr);
@@ -131,7 +131,7 @@ void MoreIconsAPI::updateRobotSprite(GJRobotSprite* sprite, const std::string& i
 void MoreIconsAPI::updatePlayerObject(PlayerObject* object, const std::string& icon, IconType type) {
     if (!object || icon.empty() || !hasIcon(icon, type)) return;
 
-    MoreIconsAPI::setUserObject(object, icon);
+    setUserObject(object, icon);
 
     if (type == IconType::Robot || type == IconType::Spider) {
         auto robotSprite = type == IconType::Robot ? object->m_robotSprite : object->m_spiderSprite;