+++ src/hooks/AppDelegate.cpp
@@ -9,10 +9,30 @@ class $modify(MIAppDelegate, AppDelegate) {
     }
 
     void trySaveGame(bool closing) {
-        MoreIcons::removeSaved();
+        auto mod = Mod::get();
+
+        mod->setSavedValue<std::vector<std::string>>("icons", {});
+        mod->setSavedValue<std::vector<std::string>>("ships", {});
+        mod->setSavedValue<std::vector<std::string>>("balls", {});
+        mod->setSavedValue<std::vector<std::string>>("ufos", {});
+        mod->setSavedValue<std::vector<std::string>>("waves", {});
+        mod->setSavedValue<std::vector<std::string>>("robots", {});
+        mod->setSavedValue<std::vector<std::string>>("spiders", {});
+        mod->setSavedValue<std::vector<std::string>>("swings", {});
+        mod->setSavedValue<std::vector<std::string>>("jetpacks", {});
+        mod->setSavedValue<std::vector<std::string>>("trails", {});
 
         AppDelegate::trySaveGame(closing);
 
-        MoreIcons::restoreSaved();
+        mod->setSavedValue("icons", MoreIconsAPI::ICONS);
+        mod->setSavedValue("ships", MoreIconsAPI::SHIPS);
+        mod->setSavedValue("balls", MoreIconsAPI::BALLS);
+        mod->setSavedValue("ufos", MoreIconsAPI::UFOS);
+        mod->setSavedValue("waves", MoreIconsAPI::WAVES);
+        mod->setSavedValue("robots", MoreIconsAPI::ROBOTS);
+        mod->setSavedValue("spiders", MoreIconsAPI::SPIDERS);
+        mod->setSavedValue("swings", MoreIconsAPI::SWINGS);
+        mod->setSavedValue("jetpacks", MoreIconsAPI::JETPACKS);
+        mod->setSavedValue("trails", MoreIconsAPI::TRAILS);
     }
 };